"Fix" FTBFS by disabling source incompatibility checks
Currently, we're missing the additional lzma_ret values introduced
in 5.4 (which this copy of the library will never return). We're
also missing a number of renamed reserved fields (the struct sizes
have not changed, so the ABI is fine, but any application filling
these fields in and expecting them to be considered will be sorely
disappointed).